+++ tests/helpers.test.ts
@@ -12,11 +12,14 @@ describe('helpers', () => {
       expect(toDataString(undefined)).to.deep.equal(`{}`); // TODO: force undefined
     });
 
-    it('toDataString() works on structures', () => {
+    it('toDataString() works on basic structures', () => {
       // may have issues with formatting
-      expect(toDataString(new Date(1657196329497))).to.deep.equal(`{"__data__":"2022-07-07T12:18:49.497Z"}`);
       expect(toDataString({ foo: 'bar' })).to.deep.equal(`{"__data__":{"foo":"bar"}}`);
       expect(toDataString([1, 'foo', { val: 3 }])).to.deep.equal(`{"__data__":[1,"foo",{"val":3}]}`);
+    });
+    it('toDataString() works on hydrateable structures', () => {
+      // may have issues with formatting
+      expect(toDataString(new Date(1657196329497))).to.deep.equal(`{"__data__":"2022-07-07T12:18:49.497Z","__type__":"date"}`);
     });
   });
 
@@ -34,6 +37,24 @@ describe('helpers', () => {
       expect(fromDataString(`{"__data__":"2022-07-07T12:18:49.497Z"}`)).to.deep.equal('2022-07-07T12:18:49.497Z');
       expect(fromDataString(`{"__data__":{"foo":"bar"}}`)).to.deep.equal({ foo: 'bar' });
       expect(fromDataString(`{"__data__":[1,"foo",{"val":3}]}`)).to.deep.equal([1, 'foo', { val: 3 }]);
+    });
+
+    it('fromDataString() rehydrates dates when conditions are met', () => {
+      // Config not enabled
+      expect(fromDataString(`{"__data__":"2022-07-07T12:18:49.497Z","__type__":"date"}`)).to.deep.equal('2022-07-07T12:18:49.497Z');
+
+      // Config enabled, but type not set
+      expect(fromDataString(`{"__data__":"2022-07-07T12:18:49.497Z"}`, {
+        transforms: {
+          'Date': true
+        },
+      })).to.deep.equal('2022-07-07T12:18:49.497Z');
+
+      expect(fromDataString(`{"__data__":"2022-07-07T12:18:49.497Z","__type__":"date"}`, {
+        transforms: {
+          'Date': true
+        },
+      })).to.deep.equal(new Date(1657196329497));
     });
   });
 });